move swipe.py to core
