+++ nuvox_algorithm/core/__init__.py
@@ -1,3 +1,4 @@
 from .key import Key
 from .keyboard import Keyboard
 from .nuvox_keyboard import nuvox_keyboard
+from .swipe import Swipe, TracePoint
