+++ nuvox_algorithm/core/swipe.py
