+++ nuvox_algorithm/trace_algorithm/__init__.py
@@ -1,3 +1,1 @@
 from .trace_algorithm import TraceAlgorithm
-from .swipe import Swipe, TracePoint
-
