+++ nuvox_algorithm/trace_algorithm/utils/__init__.py
@@ -1,4 +1,5 @@
 from .download_dataset import download_trace_algorithm_train_set, download_trace_algorithm_test_set
 from .load_dataset import load_train_set, load_test_set
-
+from .rdp import rdp
+from .angle import angle
